pyproject.toml
[project]
name = "sentiric-stt-whisper-service"
version = "1.0.0"
description = "Expert Whisper transcription service for Sentiric ecosystem"
authors = [
    {name = "Sentiric Team"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core Framework
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "httpx==0.27.0",
    
    # STT Engine
    "faster-whisper==1.0.3",
    "ctranslate2==4.3.1",
    "torch==2.3.0",
    
    # Audio Processing
    "soundfile==0.12.1",
    "librosa==0.10.1", 
    "numpy==1.26.4",
    "python-multipart==0.0.6",
    "pydub==0.25.1",
    
    # Configuration & Logging
    "pydantic==2.12.3",
    "pydantic-settings==2.4.0",
    "structlog==23.3.0",
    
    # Monitoring
    "prometheus-client==0.20.0",
    "prometheus-fastapi-instrumentator==6.0.0",
    
    # gRPC & Contracts - YENİ EKLENDİ
    "grpcio>=1.62.0",
    "grpcio-tools>=1.62.0",
    "sentiric-contracts-py @ git+https://github.com/sentiric/sentiric-contracts.git@v1.9.4",
    
    # System & Utilities
    "aiofiles==23.2.1",
    "psutil==5.9.8",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.3.0",
]

dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.5",
    "pytest-cov==4.1.0",
    "black==23.12.1", 
    "flake8==7.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"